Added GraphQL fragments. Clean up.
