+++ packages/ckeditor5-dev-stale-bot/bin/utils/prepareoptions.js
