+++ packages/ckeditor5-dev-stale-bot/lib/graphql/searchissuestostale.graphql
@@ -1,3 +1,30 @@
+fragment labeledEvent on LabeledEvent {
+	createdAt
+	actor {
+		login
+	}
+	label {
+		name
+	}
+}
+
+fragment unlabeledEvent on UnlabeledEvent {
+	createdAt
+	actor {
+		login
+	}
+	label {
+		name
+	}
+}
+
+fragment issueComment on IssueComment {
+	updatedAt
+	author {
+		login
+	}
+}
+
 query SearchIssuesToStale($query: String!, $cursor: String) {
 	search(type: ISSUE, query: $query, first: 100, after: $cursor) {
 		issueCount
@@ -14,30 +41,9 @@ query SearchIssuesToStale($query: String!, $cursor: String) {
 				}
 				timelineItems(first: 100, itemTypes: [LABELED_EVENT, UNLABELED_EVENT, ISSUE_COMMENT]) {
 					nodes {
-						... on LabeledEvent {
-							createdAt
-							actor {
-								login
-							}
-							label {
-								name
-							}
-						}
-						... on UnlabeledEvent {
-							createdAt
-							actor {
-								login
-							}
-							label {
-								name
-							}
-						}
-						... on IssueComment {
-							updatedAt
-							author {
-								login
-							}
-						}
+						...labeledEvent
+						...unlabeledEvent
+						...issueComment
 					}
 					pageInfo {
 						cursor: endCursor
@@ -57,30 +63,9 @@ query SearchIssuesToStale($query: String!, $cursor: String) {
 				}
 				timelineItems(first: 100, itemTypes: [LABELED_EVENT, UNLABELED_EVENT, ISSUE_COMMENT]) {
 					nodes {
-						... on LabeledEvent {
-							createdAt
-							actor {
-								login
-							}
-							label {
-								name
-							}
-						}
-						... on UnlabeledEvent {
-							createdAt
-							actor {
-								login
-							}
-							label {
-								name
-							}
-						}
-						... on IssueComment {
-							updatedAt
-							author {
-								login
-							}
-						}
+						...labeledEvent
+						...unlabeledEvent
+						...issueComment
 					}
 					pageInfo {
 						cursor: endCursor
